Bundle our own copy of pymodbus.
This makes us independent of the version HA brings natively via its
component/modbus. pymodus does not use semver so minor version increments
may well break the API. Recently, HA bumped pymodbus to 3.9.2 and there
are patches in the pipeline to move to 3.11.1+ which will definitely
break our way of usage. There are no easy ways in Python to be compatible
to two different APIs, so we will bundle 3.9.2 with this commit.

There are other components doing this as well, for similar reasons.x